+++ src/main/java/com/aston/userservice/domain/response/UserRequisitesResponseDTO.java
@@ -2,9 +2,11 @@ package com.aston.userservice.domain.response;
 
 import lombok.Builder;
 
+import java.math.BigInteger;
+
 /**
  * DTO для направления необходимых реквизитов пользователя
  */
 @Builder
-public record UserRequisitesResponseDTO(String firstName, Integer accountNumber, Integer kbk) {
+public record UserRequisitesResponseDTO(String firstName, BigInteger accountNumber, BigInteger kbk) {
 }
